microservices/services/products/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.7.1
python-jose==3.3.0
python-multipart==0.0.9

